fix: migrate test files to ava
