+++ packages/js/__tests__/account-manager.ava.ts
@@ -1,27 +1,25 @@
+import {ava as test} from 'near-runner-ava';
 import {AccountManager, TestnetManager, TestnetRuntime, Runner} from '..';
 
-jest.setTimeout(300_000);
-
-describe('Account Manager', () => {
-  if (!Runner.networkIsTestnet()) {
-    test('skipping on ' + Runner.getNetworkFromEnv(), () => {});
-    return;
-  }
-
-  test('should create a new account', async () => {
+if (Runner.networkIsTestnet()) {
+  test('should create a new account', async t => {
     const accountManager = AccountManager.create(TestnetRuntime.defaultConfig);
     await accountManager.init();
     const {root} = accountManager;
-    expect(await root.exists()).toBe(true);
+    t.true(await root.exists());
   });
 
-  test('should be able to add funds', async () => {
+  test('should be able to add funds', async t => {
     const accountManager = AccountManager.create(TestnetRuntime.defaultConfig);
     await accountManager.init();
     const {root} = accountManager;
     const balance = await root.availableBalance();
     await (accountManager as TestnetManager).addFundsFromNetwork();
     const newBalance = await root.availableBalance();
-    expect(balance.lt(newBalance)).toBe(true);
+    t.true(balance.lt(newBalance));
   });
-});
+} else {
+  test('skipping on ' + Runner.getNetworkFromEnv(), t => {
+    t.true(true);
+  });
+}
